double -> float

+++ useful/math/math.cpp
@@ -1,6 +1,6 @@
 #include"math.h"
 
-double math::isZero(const double& number, const int& acc, const int& coef){
+float math::isZero(const float& number, const int& acc, const int& coef){
         if (std::abs(number) < coef * pow(10.0, acc)){
             return 0;
         }
